requirements.txt
aiogram==2.0.1
aiohttp==3.5.4
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
Babel==2.6.0
bcrypt==3.1.6
beautifulsoup4==4.7.1
boto==2.49.0
boto3==1.9.132
botocore==1.12.132
bz2file==0.98
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
cssselect==1.0.3
decorator==4.4.0
docutils==0.14
feedfinder2==0.0.4
feedparser==5.2.1
gensim==3.7.2
gensim-sum-ext==0.1.2
idna==2.8
jieba3k==0.35.1
jmespath==0.9.4
lxml==4.3.3
motor==2.0.0
multidict==4.5.2
networkx==2.3
newspaper3k==0.2.8
nltk==3.4.1
numpy==1.16.2
paramiko==2.4.2
Pillow==6.0.0
pyasn1==0.4.5
pycorenlp==0.3.0
pycparser==2.19
pymongo==3.7.2
PyNaCl==1.3.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
requests==2.21.0
requests-file==1.4.3
s3transfer==0.2.0
scipy==1.2.1
sentry-sdk==0.20.3
six==1.12.0
smart-open==1.8.2
soupsieve==1.9.1
tinysegmenter==0.3
tldextract==2.2.1
urllib3==1.24.2
yarl==1.3.0
